Refactor: follow 도메인 response/request dto 패키지 수정
