+++ src/main/java/com/divide/follow/FollowController.java
@@ -4,11 +4,11 @@ import com.divide.exception.RestApiException;
 import com.divide.exception.code.UserErrorCode;
 import com.divide.follow.dto.request.DeleteFollowRequest;
 import com.divide.follow.dto.request.GetFollowOtherRequest;
-import com.divide.follow.dto.request.GetFollowResponse;
+import com.divide.follow.dto.response.GetFollowResponse;
 import com.divide.follow.dto.request.PostFollowRequest;
 import com.divide.follow.dto.response.DeleteFollowResponse;
 import com.divide.follow.dto.response.GetFollowOtherResponse;
-import com.divide.follow.dto.response.GetFollowRequest;
+import com.divide.follow.dto.request.GetFollowRequest;
 import com.divide.follow.dto.response.PostFollowResponse;
 import com.divide.post.dto.response.Result;
 import com.divide.user.User;
