+++ src/main/java/com/divide/follow/dto/response/GetFollowResponse.java
@@ -1,4 +1,4 @@
-package com.divide.follow.dto.request;
+package com.divide.follow.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
