+++ src/main/java/com/divide/follow/dto/response/GetFollowResponseWithRelation.java
@@ -1,4 +1,4 @@
-package com.divide.follow.dto.request;
+package com.divide.follow.dto.response;
 
 import lombok.Getter;
